refactor grid to support multiple size breakpoints
